Start work on new forking+joining API
The old design has a major deficiency, in that forked computations always run to
completion independent of the originating computation, so the main thread can't
use the result of forks.  This new API provides a 'fork' primitive which returns
a computation which will block on the created fork's completion. There is a
discussion of the design ideas at the top of S.H.Conc.

A major feature of this API is that while it isn't a monad transformer, it
retains non-dependence on the IO monad - the resulting computations can be in
any monad.

I've also taken the opportunity to support multiple types of requests and tying
their type to the type of the results.  All this results in a more subtle
implementation but is much more useful.

This changeset just sets out the API and some supporting types - implementation
coming soon...